# Conflicts:
#	crates/db/src/models/coding_agent_turn.rs
#	crates/db/src/models/scratch.rs

+++ crates/db/src/models/scratch.rs
@@ -208,7 +208,8 @@ pub struct ProjectStatusConfigData {
     pub color: String,
     #[serde(default)]
     pub hidden: bool,
-    pub sort_order: i32,
+    #[ts(type = "number")]
+    pub sort_order: i64,
 }
 
 /// Data for project repo defaults scratch (default repos/branches per project)
@@ -245,6 +246,7 @@ pub struct DraftIssueData {
 
 /// The payload of a scratch, tagged by type. The type is part of the composite primary key.
 /// Data is stored as markdown string.
+#[allow(clippy::large_enum_variant)]
 #[derive(Debug, Clone, Serialize, Deserialize, TS, EnumDiscriminants)]
 #[serde(tag = "type", content = "data", rename_all = "SCREAMING_SNAKE_CASE")]
 #[strum_discriminants(name(ScratchType))]
